CaloMonitoring: Fix compilation warning.

See merge request atlas/athena!10910

Former-commit-id: 08eebcb2fd10c9492e288b27ba5e71d389805810

+++ Calorimeter/CaloMonitoring/src/LArCellMonTool.cxx
@@ -1345,8 +1345,8 @@ void LArCellMonTool::divideByOccupancy(TH2F* fraction, const TH2F* total, const 
 
   fraction->Reset();
   const size_t nCells=total->GetNcells();
-  assert(nCells==fraction->GetNcells());
-  assert(nCells==occupancy->GetNcells());
+  assert(static_cast<int>(nCells)==fraction->GetNcells());
+  assert(static_cast<int>(nCells)==occupancy->GetNcells());
  
   for (size_t i=0;i<nCells;++i) {
     const double t = total->GetBinContent(i);
